Refactor to CRE test directory example
